requirements.txt
matplotlib==3.5.2
numpy==1.22.4
opencv_python==4.5.3.56
Pillow==10.0.0
PyYAML==6.0
scikit_learn==1.3.0
scipy==1.8.1
timm==0.6.7
torchsummary==1.5.1
tqdm==4.64.0
ujson==1.35
